Move renderer from theme to web crate
