+++ crates/brace-theme/Cargo.toml
@@ -13,18 +13,14 @@ name = "brace_theme"
 path = "src/lib/lib.rs"
 
 [dependencies]
-actix = "0.8.0-alpha.2"
 brace-cli = { path = "../brace-cli" }
 brace-config = { path = "../brace-config" }
 failure = "0.1"
 path-absolutize = "1.1"
 serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-tera = "1.0.0-beta.3"
 toml = "0.5"
 url = "1.7"
 
 [dev-dependencies]
 assert_cmd = "0.11"
 tempfile = "3.0"
-futures = "0.1"
