+++ crates/brace-web-auth/src/lib/route/web/create.rs
@@ -2,8 +2,8 @@ use actix_web::error::{Error, ErrorForbidden, ErrorInternalServerError};
 use actix_web::web::{Data, Form};
 use actix_web::HttpResponse;
 use brace_db::Database;
-use brace_theme::renderer::{Renderer, Template};
 use brace_web::redirect::HttpRedirect;
+use brace_web::render::{Renderer, Template};
 use futures::future::{err, Either, Future};
 use serde_json::json;
