+++ crates/brace-web-page/src/lib/route/web/locate.rs
@@ -2,7 +2,7 @@ use actix_web::error::{Error, ErrorInternalServerError};
 use actix_web::web::Data;
 use actix_web::{HttpRequest, HttpResponse};
 use brace_db::Database;
-use brace_theme::renderer::{Renderer, Template};
+use brace_web::render::{Renderer, Template};
 use futures::future::Future;
 use serde_json::json;
