+++ crates/brace-web/src/lib/lib.rs
@@ -1,2 +1,3 @@
 pub mod config;
 pub mod redirect;
+pub mod render;
